+++ lib/frp/src/lib.rs
@@ -35,224 +35,6 @@ use std::borrow::Cow;
 
 
 
-// ===============
-// === DynNode ===
-// ===============
-
-/// Type level association between FRP data and dynamic node type. The associated `DynNode` type can
-/// differ depending on whether it is an event or behavior node, as they provide different APIs.
-/// For example, behaviors allow lookup for the current value, which does not make sense in case
-/// of events.
-pub trait KnownDynNode {
-    /// The node storage type.
-    type DynNode: DynNodeBounds + CloneRef;
-}
-
-/// Accessor.
-pub type DynNode<T> = <T as KnownDynNode>::DynNode;
-
-alias! {
-    /// Bounds required for every node.
-    DynNodeBounds = {
-        Debug + GraphvizBuilder + HasId + HasDisplayId + HasInputs + HasLabel + KnownOutput
-    }
-}
-
-// === EventDynNode ===
-
-/// Newtype wrapper for any event node.
-#[derive(Debug,Derivative,Shrinkwrap)]
-#[derivative(Clone(bound=""))]
-pub struct EventDynNode<Out> {
-    rc: Rc<dyn EventDynNodeBounds<Output=EventData<Out>>>,
-}
-
-alias! {
-    /// Bounds for any event node.
-    EventDynNodeBounds = { DynNodeBounds + HasEventTargets + EventEmitter }
-}
-
-impl<Out:Value> KnownDynNode for EventData<Out> {
-    type DynNode = EventDynNode<Out>;
-}
-
-impl<Out> Unwrap     for EventDynNode<Out> {}
-impl<Out> CloneRef   for EventDynNode<Out> {}
-impl<Out> HasContent for EventDynNode<Out> {
-    // TODO: Simplify after fixing https://github.com/rust-lang/rust/issues/68776
-    type Content = <EventDynNode<Out> as Deref>::Target;
-}
-
-impl<Out:Value> KnownOutput for EventDynNode<Out> {
-    type Output = EventData<Out>;
-}
-
-
-// === BehaviorDynNodeBounds ===
-
-/// Newtype wrapper for any behavior node.
-#[derive(Debug,Derivative,Shrinkwrap)]
-#[derivative(Clone(bound=""))]
-pub struct BehaviorDynNode<Out> {
-    rc: Rc<dyn BehaviorDynNodeBounds<Output=BehaviorData<Out>>>,
-}
-
-alias! {
-    /// Bounds for any behavior node.
-    BehaviorDynNodeBounds = { DynNodeBounds + HasCurrentValue  }
-}
-
-impl<Out:Value> KnownDynNode for BehaviorData<Out> {
-    type DynNode = BehaviorDynNode<Out>;
-}
-
-impl<Out> Unwrap     for BehaviorDynNode<Out> {}
-impl<Out> CloneRef   for BehaviorDynNode<Out> {}
-impl<Out> HasContent for BehaviorDynNode<Out> {
-    // TODO: Simplify after fixing https://github.com/rust-lang/rust/issues/68776
-    type Content = <BehaviorDynNode<Out> as Deref>::Target;
-}
-
-impl<Out:Value> KnownOutput for BehaviorDynNode<Out> {
-    type Output = BehaviorData<Out>;
-}
-
-
-
-// ============
-// === Node ===
-// ============
-
-// === Types ===
-
-/// The type of any FRP node which produces event messages. Having a reference to a node is like
-/// having a reference to network endpoint which transmits messages of a given type. Thus, it is a
-/// nice mental simplification to think about it just like about an event (stream).
-pub type Event<T> = Node<EventData<T>>;
-
-/// The type of any FRP node which can be queried for behavior value. Having a reference to a node
-/// is like having a reference to network endpoint which transmits messages of a given type. Thus,
-/// it is a nice mental simplification to think about it just like about a behavior.
-pub type Behavior <T> = Node<BehaviorData<T>>;
-
-
-// === Definition ===
-
-/// Node is used as a common types for frp operations. For example, `Event<T>` is just an alias to
-/// `Node<EventData<T>>`.
-#[derive(Derivative)]
-#[derivative(Clone(bound=""))]
-#[derivative(Debug(bound=""))]
-pub struct Node<Out:KnownDynNode> {
-    storage: DynNode<Out>,
-}
-
-impl<Out:Data> Node<Out> {
-    /// Constructor.
-    pub fn new(storage:DynNode<Out>) -> Self {
-        Self {storage}
-    }
-}
-
-
-// === Type Deps ===
-
-impl<Out:Data> KnownOutput for Node<Out> { type Output  = Out; }
-impl<Out:Data> HasContent  for Node<Out> { type Content = DynNode<Out>; }
-impl<Out:Data> Unwrap      for Node<Out> {}
-
-
-// === Instances ===
-
-impl<Out:Data> Deref for Node<Out> {
-    type Target = DynNode<Out>;
-    fn deref(&self) -> &Self::Target {
-        &self.storage
-    }
-}
-
-impl<Out:Data> CloneRef for Node<Out> {
-    fn clone_ref(&self) -> Self {
-        let storage = self.storage.clone_ref();
-        Self {storage}
-    }
-}
-
-
-// === Construction ===
-
-impl<Out:Data> From<&Node<Out>> for Node<Out> {
-    fn from(t:&Node<Out>) -> Self {
-        t.clone_ref()
-    }
-}
-
-impl<Storage,Out:Value>
-From<&Storage> for Behavior<Out>
-    where Storage : BehaviorDynNodeBounds<Output=BehaviorData<Out>> + Clone + 'static {
-    fn from(storage:&Storage) -> Self {
-        Self::new(BehaviorDynNode{rc:Rc::new(storage.clone())})
-    }
-}
-
-impl<Storage,Out:Value>
-From<&Storage> for Event<Out>
-    where Storage : EventDynNodeBounds<Output=EventData<Out>> + Clone + 'static {
-    fn from(storage:&Storage) -> Self {
-        Self::new(EventDynNode{rc:Rc::new(storage.clone())})
-    }
-}
-
-
-// === AddTarget ===
-
-/// Abstraction for adding a target to a given node. Nodes which carry behaviors do not need to
-/// perform any operation here, while event streams want to register the nodes they want to send
-/// notifications to.
-pub trait AddTarget<T> {
-    /// Adds a node as a target of the current flow.
-    fn add_target(&self,t:&T);
-}
-
-impl<S,T:Value> AddTarget<S> for Event<T>
-    where for<'t> &'t S : Into<AnyEventConsumer<EventData<T>>> {
-    fn add_target(&self,t:&S) {
-        self.add_event_target(t.into())
-    }
-}
-
-impl<S,T:Value> AddTarget<S> for Behavior<T> {
-    fn add_target(&self,_:&S) {}
-}
-
-
-
-// ===============
-// === AnyNode ===
-// ===============
-
-#[derive(Debug,Shrinkwrap)]
-pub struct AnyNode {
-    rc: Rc<dyn AnyNodeOps>,
-}
-
-alias! { AnyNodeOps = { Debug + GraphvizBuilder + HasId + HasDisplayId + KnownOutputType } }
-
-
-// === Instances ===
-
-impls! { [Out:Data+'static] From <&Node<Out>> for AnyNode { |t| t.clone_ref().into() } }
-impls! { [Out:Data+'static] From  <Node<Out>> for AnyNode { |t| Self {rc:Rc::new(t)} } }
-
-impl KnownOutputType for AnyNode {
-    fn output_type(&self) -> DataType {
-        self.rc.output_type()
-    }
-
-    fn output_type_value_name(&self) -> String {
-        self.rc.output_type_value_name()
-    }
-}
 
 
 
@@ -415,63 +197,7 @@ fn base_type_name<T>() -> String {
 
 
 
-// =====================
-// === EventConsumer ===
-// =====================
 
-// === Definition ===
-
-/// Abstraction for nodes which are able to consume events.
-pub trait EventConsumer: KnownEventInput + Debug {
-    /// Function called on every new received event.
-    fn on_event(&self, input:&Self::EventInput);
-}
-
-
-// === AnyEventConsumer ===
-
-/// Abstraction for any node which consumes events of a given type.
-#[derive(Clone,Debug,Shrinkwrap)]
-pub struct AnyEventConsumer<In> {
-    raw: Rc<dyn EventConsumer<EventInput=In>>,
-}
-
-impl<In:Data> AnyEventConsumer<In> {
-    /// Constructor.
-    pub fn new<A:EventConsumer<EventInput=In>+'static>(a:A) -> Self {
-        let raw = Rc::new(a);
-        Self {raw}
-    }
-}
-
-impl<T,In> From<&T> for AnyEventConsumer<In>
-    where T  : EventConsumer<EventInput=In> + Clone + 'static,
-          In : Data {
-    fn from(t:&T) -> Self {
-        Self::new(t.clone())
-    }
-}
-
-
-
-// ====================
-// === EventEmitter ===
-// ====================
-
-// === Definition ===
-
-/// Abstraction for nodes which are able to consume events.
-pub trait EventEmitter: KnownOutput {
-    /// Function called on every new received event.
-    fn emit(&self, event:&Self::Output);
-}
-
-impl<T> EventEmitter for T
-where T:Unwrap+KnownOutput, Content<T>:EventEmitter<Output=Output<Self>> {
-    fn emit(&self, event:&Self::Output) {
-        self.unwrap().emit(event)
-    }
-}
 
 
 
@@ -528,9 +254,9 @@ inference_rules! {
 
 
 
-// ============================
+// =========================
 // === ContainsEventData ===
-// ============================
+// =========================
 
 pub trait ContainsEventData {
     type Result : Data;
@@ -539,35 +265,59 @@ pub trait ContainsEventData {
 pub type SelectEventData<T> = <T as ContainsEventData>::Result;
 
 impl<T1> ContainsEventData for EventData<T1>
-    where EventData<T1> : Data {
+where EventData<T1> : Data {
+    type Result = EventData<T1>;
+}
+
+impl<T1,T2> ContainsEventData for (EventData<T1>,EventData<T2>)
+where EventData<T1> : Data {
     type Result = EventData<T1>;
 }
 
 impl<T1,T2> ContainsEventData for (EventData<T1>,BehaviorData<T2>)
-    where EventData<T1> : Data {
+where EventData<T1> : Data {
     type Result = EventData<T1>;
 }
 
 impl<T1,T2> ContainsEventData for (BehaviorData<T1>,EventData<T2>)
-    where EventData<T2> : Data {
+where EventData<T2> : Data {
     type Result = EventData<T2>;
 }
 
 impl<T1,T2,T3> ContainsEventData for (EventData<T1>,BehaviorData<T2>,BehaviorData<T3>)
-    where EventData<T1> : Data {
+where EventData<T1> : Data {
     type Result = EventData<T1>;
 }
 
 impl<T1,T2,T3> ContainsEventData for (BehaviorData<T1>,EventData<T2>,BehaviorData<T3>)
-    where EventData<T2> : Data {
+where EventData<T2> : Data {
     type Result = EventData<T2>;
 }
 
 impl<T1,T2,T3> ContainsEventData for (BehaviorData<T1>,BehaviorData<T2>,EventData<T3>)
-    where EventData<T3> : Data {
+where EventData<T3> : Data {
     type Result = EventData<T3>;
 }
 
+impl<T1,T2,T3> ContainsEventData for (EventData<T1>,EventData<T2>,BehaviorData<T3>)
+where EventData<T1> : Data {
+    type Result = EventData<T1>;
+}
+
+impl<T1,T2,T3> ContainsEventData for (EventData<T1>,BehaviorData<T2>,EventData<T3>)
+    where EventData<T1> : Data {
+    type Result = EventData<T1>;
+}
+
+impl<T1,T2,T3> ContainsEventData for (BehaviorData<T1>,EventData<T2>,EventData<T3>)
+where EventData<T2> : Data {
+    type Result = EventData<T2>;
+}
+
+impl<T1,T2,T3> ContainsEventData for (EventData<T1>,EventData<T2>,EventData<T3>)
+where EventData<T1> : Data {
+    type Result = EventData<T1>;
+}
 
 
 // =================================================================================================
@@ -634,7 +384,7 @@ impl<Out:KnownSourceStorage> HasInputs for SourceShape<Out> {
 
 
 
-macro_rules! define_node {
+macro_rules! define_X_node {
     (
         $(#$meta:tt)*
         pub struct $name:ident $shape_name:ident [$($poly_input:ident)*]
@@ -666,32 +416,78 @@ macro_rules! define_node {
 }
 
 
+macro_rules! define_node {
+    (
+        $(#$meta:tt)*
+        $name:ident $shape_name:ident [$($poly_input:ident)*] -> [$($out:tt)*]
+            { $( $field:ident : $field_type:ty ),* }
+    ) => {
+        $(#$meta)*
+        pub type $name<$($poly_input,)*> = NodeWrapper<$shape_name<$($poly_input,)*>>;
+
+        $(#$meta)*
+        #[derive(Debug)]
+        #[allow(non_camel_case_types)]
+        pub struct $shape_name<$($poly_input:Data,)*> {
+            $( $poly_input : Node<$poly_input> ),* ,
+            $( $field      : $field_type ),*
+        }
+
+        impl<$($poly_input:Data,)*>
+        KnownOutput for $shape_name<$($poly_input,)*> {
+            type Output = $($out)*;
+        }
+
+        impl<$($poly_input:Data,)*>
+        KnownEventInput for $shape_name<$($poly_input,)*>
+        where ($($poly_input),*) : ContainsEventData,
+              SelectEventData<($($poly_input),*)> : Data {
+            type EventInput = SelectEventData<($($poly_input),*)>;
+        }
+
+//        impl<$($poly_input:Data,)*> $name<$($poly_input,)*>
+//        where Self:KnownOutput , $(Node<$poly_input> : AddTarget<Self>,)* {
+//            fn new_named<Label,$($poly_input:Data,)*>
+//            (label:Label, $($poly_input:$poly_input,)*) -> Self {
+//                let $poly_input = $poly_input.into();
+//                let input_ref   = $poly_input.clone();
+//            }
+//        }
+
+    }
+}
+
+
 
 // =============
 // === Merge ===
 // =============
 
-pub type Merge<T> = NodeWrapper<MergeShape<T>>;
-
-#[derive(Debug)]
-#[allow(non_camel_case_types)]
-pub struct MergeShape<T:Data> {
-    source1 : Node<T>,
-    source2 : Node<T>,
+define_node! {
+    Merge MergeShape [source1 source2] -> [source1] {}
 }
 
-impl<T:Data> KnownOutput     for MergeShape<T> { type Output     = T; }
-impl<T:Data> KnownEventInput for MergeShape<T> { type EventInput = T; }
+//pub type Merge<T1,T2> = NodeWrapper<MergeShape<T1,T2>>;
+//
+//#[derive(Debug)]
+//#[allow(non_camel_case_types)]
+//pub struct MergeShape<T1:Data,T2:Data> {
+//    source1 : Node<T1>,
+//    source2 : Node<T2>,
+//}
+
+//impl<T:Data> KnownOutput     for MergeShape<T,T> { type Output     = T; }
+//impl<T:Data> KnownEventInput for MergeShape<T,T> { type EventInput = T; }
 
 
 // === Constructor ===
 
-impl<T:Data> Merge<T>
-    where Node<T> : AddTarget<Self> {
+impl<T1:Data,T2:Data> Merge<T1,T2>
+where Node<T1> : AddTarget<Self>, Node<T2> : AddTarget<Self>, MergeShape<T1,T2>: KnownOutput {
     fn new_named<Label,Source1,Source2>(label:Label, source1:Source1, source2:Source2) -> Self
         where Label   : Into<CowString>,
-              Source1 : Into<Node<T>>,
-              Source2 : Into<Node<T>> {
+              Source1 : Into<Node<T1>>,
+              Source2 : Into<Node<T2>> {
         let source1     = source1.into();
         let source2     = source2.into();
         let source1_ref = source1.clone();
@@ -703,13 +499,20 @@ impl<T:Data> Merge<T>
     }
 }
 
-impl<T:Value> EventConsumer for Merge<EventData<T>> {
+impl<T1:Data,T2:Data> EventConsumer for Merge<T1,T2>
+where MergeShape<T1,T2> : KnownEventInput<EventInput=Output<Self>>
+{
     fn on_event(&self, event:&Self::EventInput) {
         self.emit_event(event);
     }
 }
 
-impl<T:Data> HasInputs for MergeShape<T> {
+//pub trait EventConsumer: KnownEventInput + Debug {
+//    /// Function called on every new received event.
+//    fn on_event(&self, input:&Self::EventInput);
+//}
+
+impl<T:Data> HasInputs for MergeShape<T,T> {
     fn inputs(&self) -> Vec<AnyNode> {
         vec![(&self.source1).into(), (&self.source2).into()]
     }
@@ -1034,7 +837,7 @@ impl<In1:Data, In2:Data> HasInputs for GateShape<In1,In2> {
 // === Lambda ===
 // ==============
 
-define_node! {
+define_X_node! {
     /// Transforms input data with the provided function. Lambda accepts a single input and outputs
     /// message of the same type as the input message.
     pub struct Lambda LambdaShape [source] {
@@ -1122,7 +925,7 @@ impl<In1:Data, Out:Data> HasInputs for LambdaShape<In1,Out> {
 // === Lambda2 ===
 // ===============
 
-define_node! {
+define_X_node! {
     /// Transforms input data with the provided function. `Lambda2` accepts two inputs. If at least
     /// one of the inputs was event, the output message will be event as well. In case both inputs
     /// were behavior, a new behavior will be produced.
