+++ lib/frp/src/node.rs
@@ -1,5 +1,7 @@
+pub mod class;
 pub mod id;
 pub mod label;
 
+pub use class::*;
 pub use id::*;
 pub use label::*;
